recog_system/pyproject.toml
[project]
name = "recog-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.6",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "timm>=1.0.12",
    "torch>=2.10.0",
    "torchvision>=0.21.0",
    "transformers",
    "mediapipe>=0.10.20",
    "opencv-python>=4.11.0.86",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
    "joblib>=1.5.3",
    "numpy>=2.4.3",
    "requests>=2.32.5",
    "typing-extensions>=4.15.0",
]
